[Host.Memory] Similar to MediatR the memory bus should have ability to auto declare producers/consumers and handlers #116
