+++ docs/provider_memory.md
@@ -6,7 +6,7 @@ Please read the [Introduction](intro.md) before reading this provider documentat
 - [Configuration](#configuration)
   - [Disabling serialization](#disabling-serialization)
   - [Virtual Topics](#virtual-topics)
-  - [Autoregistration](#autoregistration)
+  - [Auto Declaration](#auto-declaration)
 - [Lifecycle](#lifecycle)
   - [Blocking Publish](#blocking-publish)
   - [Per-Message DI scope](#per-message-di-scope)
@@ -17,9 +17,7 @@ The Memory transport provider can be used for internal communication within the 
 
 > Since messages are passed in memory and never persisted, they will be lost if your app dies while consuming these messages.
 
-A fitting use case for in memory communication is to integrate your domain layer with other application layers via domain events.
-
-> The request-response has been implemented with completion of issue #7.
+Good use case for in memory communication is to integrate the domain layer with other application layers via domain events pattern.
 
 ## Configuration
 
@@ -27,7 +25,7 @@ The memory transport is configured using the `.WithProviderMemory()`:
 
 ```cs
 // MessageBusBuilder mbb;
-mbb.    
+mbb
    .WithProviderMemory(new MemoryMessageBusSettings())
    .WithSerializer(new JsonMessageSerializer());
 ```
@@ -38,7 +36,7 @@ Since messages are passed within the same process, serializing and deserializing
 
 ```cs
 // MessageBusBuilder mbb;
-mbb.    
+mbb
    .WithProviderMemory(new MemoryMessageBusSettings
    {
       // Do not serialize the domain events and rather pass the same instance across handlers (faster 
@@ -51,13 +49,13 @@ mbb.
 
 ### Virtual Topics
 
-Unlike other transport providers, memory transport does not have true notion of topics (or queues). However, it is still required to use topic names. This is required, so that the bus knows on which virtual topic to deliver the message to and from what virtual topic to consume.
+Unlike other transport providers, memory transport does not have true notion of topics (or queues). However, it is still required to use topic names. This is required, so that the bus knows on which virtual topic to deliver the message to, and from what virtual topic to consume.
 
 Here is an example for the producer side:
 
 ```cs
 // declare that OrderSubmittedEvent will be produced
-mbb.Produce<OrderSubmittedEvent>(x => x.DefaultTopic(x.Settings.MessageType.Name));
+mbb.Produce<OrderSubmittedEvent>(x => x.DefaultTopic(x.MessageType.Name));
 
 // alternatively
 mbb.Produce<OrderSubmittedEvent>(x => x.DefaultTopic("OrderSubmittedEvent"));
@@ -75,39 +73,49 @@ mbb.Consume<OrderSubmittedEvent>(x => x.Topic("OrderSubmittedEvent").WithConsume
 
 The benefit is that we can channel messages of the same type via different virtual topics.
 
-### Autoregistration
+### Auto Declaration
 
-During configuration, we can discover all the handlers (or messages that will be produced) and register them in the bus. This can be useful to autoregister all of your domain handlers from the application layer. 
+> Since 1.19.1
 
-It required a bit of reflection code and the `.Do(Action<MessageBusBuilder> action)` method:
+During bus configuration, we can leverage `AutoDeclareFromConsumers()` method to discover all the consumers (`IConsumer<T>`) and handlers (`IRequestHandler<T,R>`) types and auto declare the respective producers and consumers/handlers in the bus. This can be useful to auto declare all of the domain event handlers in an application layer.
 
 ```cs
-Assembly applicationLayer = ...
-
-// auto discover consumers/handler (classes that implement IConsumer<T>) in the applicationLayer assembly and register with the bus
-mbb.Do(builder => applicationLayer.GetTypes().Where(t => t.IsClass && !t.IsAbstract)
-                    .SelectMany(t => t.GetInterfaces(), (t, i) => new { Type = t, Interface = i })
-                    .Where(x => x.Interface.IsGenericType && x.Interface.GetGenericTypeDefinition() == typeof(IConsumer<>))
-                    .Select(x => new { HandlerType = x.Type, EventType = x.Interface.GetGenericArguments()[0] })
-                    .ToList()
-                    .ForEach(find =>
-                    {
-                        Log.InfoFormat(CultureInfo.InvariantCulture, "Registering {0} in the bus", find.EventType);
-                        builder.Produce(find.EventType, x => x.DefaultTopic(x.Settings.MessageType.Name));
-                        builder.Consume(find.EventType, x => x.Topic(x.MessageType.Name).WithConsumer(find.HandlerType));
-                    })
-                );
+mbb
+   .WithProviderMemory(new MemoryMessageBusSettings())
+   .AutoDeclareFromConsumers(Assembly.GetExecutingAssembly());
+   // If we want to filter to specific consumer/handler types then we can supply an additional filter:
+   //.AutoDeclareFromConsumers(Assembly.GetExecutingAssembly(), consumerTypeFilter: (consumerType) => consumerType.Name.EndsWith("Handler"));
 ```
+
+For example, assuming this is the discovered type:
+
+```cs
+public class EchoRequestHandler : IRequestHandler<EchoRequest, EchoResponse>
+{
+   public Task<EchoResponse> OnHandle(EchoRequest request, string name) { /* ... */ }
+}
+```
+
+The bus registrations will end up:
+
+```cs
+mbb.Produce<EchoRequest>(x => x.DefaultTopic(x.MessageType.Name));
+mbb.Handle<EchoRequest, EchoResponse>(x => x.Topic(x.MessageType.Name).WithConsumer<EchoRequestHandler>());
+```
+
+> Using `AutoDeclareFromConsumers` to configure the memory bus is recommended, as it provides a good developer experience.
 
 ## Lifecycle
 
 ### Blocking Publish
 
-The `Send<T>()` is blocking and `Publsh<T>()` is blocking by default. 
+The `Send<T>()` is blocking and `Publsh<T>()` is blocking by default.
 It might be expected that the `Publish<T>()` to be fire-and-forget and non-blocking, however this is to ensure the consumer/handler have been processed by the time the method call returns. That behavior is optimized for domain-events where you expect the side effects to be executed synchronously.
 
 ToDo: In the future we will want to expose a setting to make the `Publish<T>()` non-blocking.
 
+> In contrast to MediatR, having the `Publish<T>` blocking, allows us to avoid having to use `Send<T, R>` and have the developer to use `Unit` or `VoidResult`.
+
 ### Per-Message DI scope
 
-> Unlike stated for the [Introduction](intro.md) the memory bus has per-message scoped disabled by default.
+> Unlike stated for the [Introduction](intro.md) the memory bus has per-message scoped disabled by default. However, the memory bus consumer/handler would join the already ongoing DI scope. This is desired in scenarios where an external message is being handled as part of the unit of work (Kafka/ASB, etc) or an API HTTP request is being handled - the memory bus consumer/handler will be looked up in the ongoing/current DI scope.
